refactor: replace db into separate module
