+++ src/DB/db.module.ts
@@ -1,0 +1,9 @@
+import { Global, Module } from '@nestjs/common';
+import { DataBaseService } from './db.service';
+
+@Global()
+@Module({
+  providers: [DataBaseService],
+  exports: [DataBaseService],
+})
+export class Database {}
