+++ src/albums/albums.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 import { AlbumsService } from './albums.service';
 import { AlbumsController } from './albums.controller';
-import { AlbumsStorage } from './storage/albums.storage';
 
 @Module({
   controllers: [AlbumsController],
-  providers: [AlbumsService, AlbumsStorage],
+  providers: [AlbumsService],
 })
 export class AlbumsModule {}
