+++ src/tracks/tracks.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TracksService } from './tracks.service';
 import { TracksController } from './tracks.controller';
-import { TracksStorage } from './storage/tracks.storage';
 
 @Module({
   controllers: [TracksController],
-  providers: [TracksService, TracksStorage],
+  providers: [TracksService],
 })
 export class TracksModule {}
