+++ src/tracks/tracks.service.ts
@@ -1,29 +1,53 @@
 import { Injectable } from '@nestjs/common';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
-import { TracksStorage } from './storage/tracks.storage';
 import { TrackEntity } from './entities/track.entity';
+import { DataBaseService } from 'src/DB/db.service';
+import { NotFoundException } from '@nestjs/common/exceptions';
+import { v4 as uuidv4 } from 'uuid';
 
 @Injectable()
 export class TracksService {
-  constructor(private storage: TracksStorage) {}
-  create(createUserDto: CreateTrackDto): TrackEntity {
-    return this.storage.create(createUserDto);
+  constructor(private db: DataBaseService) {}
+  create(createTrackDto: CreateTrackDto): TrackEntity {
+    const track = {
+      ...createTrackDto,
+      id: uuidv4(),
+    };
+
+    this.db.tracks.push(track);
+    return track;
   }
 
   findAll(): TrackEntity[] {
-    return this.storage.findAll();
+    return this.db.tracks;
   }
 
   findOne(id: string): TrackEntity {
-    return this.storage.findOne(id);
+    const track = this.db.tracks.filter(
+      (track: TrackEntity) => track.id === id,
+    )[0];
+    if (track) {
+      return track;
+    }
+    throw new NotFoundException();
   }
 
-  update(id: string, updateUserDto: UpdateTrackDto) {
-    return this.storage.update(id, updateUserDto);
+  update(id: string, updateTrackDto: UpdateTrackDto) {
+    const track = this.findOne(id);
+    track.name = updateTrackDto.name;
+    track.artistId = updateTrackDto.artistId;
+    track.albumId = updateTrackDto.albumId;
+    track.duration = updateTrackDto.duration;
+    return track;
   }
 
   remove(id: string) {
-    return this.storage.remove(id);
+    const track = this.findOne(id);
+    if (track) {
+      this.db.tracks = this.db.tracks.filter(
+        (track: TrackEntity) => track.id !== id,
+      );
+    }
   }
 }
