+++ src/users/users.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
-import { UserStorage } from './storage/users.storage';
 
 @Module({
   controllers: [UsersController],
-  providers: [UsersService, UserStorage],
+  providers: [UsersService],
 })
 export class UsersModule {}
